+++ gui/Common.h
@@ -308,7 +308,7 @@ void ProcessCommandLineArgs(int argc, char *argv[]);
 
 void RunHOGGUI(int argc, char* argv[], int windowDimension = 1000);
 void RunHOGGUI(int argc, char* argv[], int xDimension, int yDimension);
-void SaveScreenshot(int windowID, const char *filename);
+void SaveScreenshot(unsigned long windowID, const char *filename);
 void SetZoom(int windowID, float amount);
 
 void submitTextToBuffer(const char *val);
